feat: update daos

+++ src/main/java/com/cyrildewit/pgc/services/UserServiceInterface.java
@@ -15,7 +15,7 @@ public interface UserServiceInterface
 {
     public List<User> getAllUsers();
 
-    public Optional<User> findUserById(Integer id);
+    public Optional<User> findUserById(Long id);
 
     public Optional<User> findUserByUuid(UUID uuid);
 
@@ -23,7 +23,7 @@ public interface UserServiceInterface
 
     public void updateUser(User user, String[] params);
 
-    public void deleteUserById(Integer id);
+    public void deleteUserById(Long id);
 
     public void deleteUser(User user);
 }
